text2mask/requirements.txt
anyio==3.6.2
asgiref==3.4.1
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
chardet==4.0.0
click==7.1.2
-e git+https://github.com/openai/CLIP.git#egg=clip
commonmark==0.9.1
contextlib2==21.6.0
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
fastapi==0.83.0
ftfy==6.0.3
h11==0.10.0
idna==2.10
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.4.0
pip-search==0.0.12
pkg-resources==0.0.0
pydantic==1.9.2
Pygments==2.14.0
pyparsing==3.0.9
python-common-service==1.0.3
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.25.1
rich==12.6.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
typing-extensions==4.1.1
urllib3==1.26.14
uvicorn==0.12.0
wcwidth==0.2.6
zipp==3.6.0
